step3 refactor buying money, winning number input 객체 DTO 네임스페이스로 변경
