+++ src/test/java/lotto/automatic/domain/LottoWinningNumsTest.java
@@ -1,5 +1,6 @@
 package lotto.automatic.domain;
 
+import lotto.automatic.dto.LottoWinningNums;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
